pyproject.toml
[project]
name = "track-sorter"
version = "0.1.0"
description = "Sort your tracks using tracklists"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "returns>=0.26.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]

[project.scripts]
track-sorter = "track_sorter:cli"

[tools.setuptools.package.find]
where = ["src"]

[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"
